skin-cancer-backend/requirements.txt
flask
flask-cors
flask-pymongo

tensorflow
keras

numpy
pillow

pyjwt
bcrypt
python-dotenv

pymongo
dnspython

email-validator
